+++ example_backend_app/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ExampleBackendAppConfig(AppConfig):
+    name = 'example_backend_app'