+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Unreleased
 Added
 -----
 
+- **General**
+    - ``example_backend_app`` for a minimal backend app example
 - **Timeline**
     - Add timeline app based on SODAR v0.3.2 (#2)
     - App documentation